(GH-9)(refactor) autouninstaller code to a service
Move the code for autouninstaller to its own service and remove the
stipulation that powershell needed to have NOT run for the
autouninstaller to run.
