+++ src/chocolatey/chocolatey.csproj
@@ -91,6 +91,7 @@
     <Compile Include="infrastructure.app\messages\PostRunEvent.cs" />
     <Compile Include="infrastructure.app\nuget\NugetPush.cs" />
     <Compile Include="infrastructure.app\runners\GenericRunner.cs" />
+    <Compile Include="infrastructure.app\services\AutomaticUninstallerService.cs" />
     <Compile Include="infrastructure.app\services\ChocolateyConfigSettingsService.cs" />
     <Compile Include="infrastructure.app\services\IChocolateyConfigSettingsService.cs" />
     <Compile Include="infrastructure.app\configuration\ConfigFileApiKeySetting.cs" />
@@ -125,6 +126,7 @@
     <Compile Include="infrastructure.app\services\IPowershellService.cs" />
     <Compile Include="infrastructure.app\services\IRegistryService.cs" />
     <Compile Include="infrastructure.app\services\IShimGenerationService.cs" />
+    <Compile Include="infrastructure.app\services\IAutomaticUninstallerService.cs" />
     <Compile Include="infrastructure.app\services\NugetService.cs" />
     <Compile Include="infrastructure.app\services\PowershellService.cs" />
     <Compile Include="infrastructure.app\domain\Registry.cs" />
