+++ src/chocolatey/infrastructure.app/services/ChocolateyPackageService.cs
@@ -24,7 +24,6 @@ namespace chocolatey.infrastructure.app.services
     using configuration;
     using domain;
     using filesystem;
-    using infrastructure.commands;
     using logging;
     using platforms;
     using results;
@@ -37,8 +36,9 @@ namespace chocolatey.infrastructure.app.services
         private readonly IFileSystem _fileSystem;
         private readonly IRegistryService _registryService;
         private readonly IChocolateyPackageInformationService _packageInfoService;
+        private readonly IAutomaticUninstallerService _autoUninstallerService;
 
-        public ChocolateyPackageService(INugetService nugetService, IPowershellService powershellService, IShimGenerationService shimgenService, IFileSystem fileSystem, IRegistryService registryService, IChocolateyPackageInformationService packageInfoService)
+        public ChocolateyPackageService(INugetService nugetService, IPowershellService powershellService, IShimGenerationService shimgenService, IFileSystem fileSystem, IRegistryService registryService, IChocolateyPackageInformationService packageInfoService, IAutomaticUninstallerService autoUninstallerService)
         {
             _nugetService = nugetService;
             _powershellService = powershellService;
@@ -46,6 +46,7 @@ namespace chocolatey.infrastructure.app.services
             _fileSystem = fileSystem;
             _registryService = registryService;
             _packageInfoService = packageInfoService;
+            _autoUninstallerService = autoUninstallerService;
         }
 
         public void list_noop(ChocolateyConfiguration config)
@@ -304,98 +305,12 @@ namespace chocolatey.infrastructure.app.services
 
                         _shimgenService.uninstall(config, packageResult);
 
-                        var powershellRan = false;
                         if (!config.SkipPackageInstallProvider)
                         {
-                            powershellRan = _powershellService.uninstall(config, packageResult);
+                             _powershellService.uninstall(config, packageResult);
                         }
 
-                        //todo run autouninstaller every time - to do this you must determine if the install path still exists.
-                        if (!powershellRan)
-                        {
-                            //todo:refactor this crap
-                            if (config.Features.AutoUninstaller)
-                            {
-                                var pkgInfo = _packageInfoService.get_package_information(packageResult.Package);
-
-                                if (pkgInfo.RegistrySnapshot != null)
-                                {
-
-                                    this.Log().Info(" Running AutoUninstaller...");
-
-                                    foreach (var key in pkgInfo.RegistrySnapshot.RegistryKeys.or_empty_list_if_null())
-                                    {
-                                        this.Log().Debug(() => " Preparing uninstall key '{0}'".format_with(key.UninstallString));
-                                        // split on " /" and " -" for quite a bit more accuracy
-                                        IList<string> uninstallArgs = key.UninstallString.to_string().Split(new[] {" /", " -"}, StringSplitOptions.RemoveEmptyEntries).ToList();
-                                        var uninstallExe = uninstallArgs.DefaultIfEmpty(string.Empty).FirstOrDefault().remove_surrounding_quotes();
-                                        this.Log().Debug(() => " Uninstaller path is '{0}'".format_with(uninstallExe));
-                                        uninstallArgs.Remove(uninstallExe);
-
-                                        if (!key.HasQuietUninstall)
-                                        {
-                                            IInstaller installer = new CustomInstaller();
-
-                                            //refactor this to elsewhere
-                                            switch (key.InstallerType)
-                                            {
-                                                case InstallerType.Msi:
-                                                    installer = new MsiInstaller();
-                                                    break;
-                                                case InstallerType.InnoSetup:
-                                                    installer = new InnoSetupInstaller();
-                                                    break;
-                                                case InstallerType.Nsis:
-                                                    installer = new NsisInstaller();
-                                                    break;
-                                                case InstallerType.InstallShield:
-                                                    installer = new CustomInstaller();
-                                                    break;
-                                                default:
-                                                    // skip
-
-                                                    break;
-                                            }
-
-                                            this.Log().Debug(() => " Installer type is '{0}'".format_with(installer.GetType().Name));
-
-                                            uninstallArgs.Add(installer.build_uninstall_command_arguments());
-                                        }
-
-                                        this.Log().Debug(() => " Args are '{0}'".format_with(uninstallArgs.join(" ")));
-
-                                        var exitCode = CommandExecutor.execute(
-                                            uninstallExe, uninstallArgs.join(" "), config.CommandExecutionTimeoutSeconds,
-                                            (s, e) =>
-                                                {
-                                                    if (string.IsNullOrWhiteSpace(e.Data)) return;
-                                                    this.Log().Debug(() => " [AutoUninstaller] {0}".format_with(e.Data));
-                                                },
-                                            (s, e) =>
-                                                {
-                                                    if (string.IsNullOrWhiteSpace(e.Data)) return;
-                                                    this.Log().Error(() => " [AutoUninstaller] {0}".format_with(e.Data));
-                                                });
-
-                                        if (exitCode != 0)
-                                        {
-                                            Environment.ExitCode = exitCode;
-                                            string logMessage = " Auto uninstaller failed. Please remove machine installation manually.";
-                                            this.Log().Error(() => logMessage);
-                                            packageResult.Messages.Add(new ResultMessage(ResultType.Error, logMessage));
-                                        }
-                                        else
-                                        {
-                                            this.Log().Info(() => " AutoUninstaller has successfully uninstalled {0} from your machine install".format_with(packageResult.Package.Id));
-                                        }
-                                    }
-                                }
-                            }
-                            else
-                            {
-                                this.Log().Info(() => "Skipping auto uninstaller due to feature not being enabled.");
-                            }
-                        }
+                        _autoUninstallerService.run(packageResult, config);
 
                         if (packageResult.Success)
                         {
