+++ src/chocolatey/infrastructure.app/services/IAutomaticUninstallerService.cs
@@ -1,0 +1,33 @@
+﻿// Copyright © 2011 - Present RealDimensions Software, LLC
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// 
+// You may obtain a copy of the License at
+// 
+// 	http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+namespace chocolatey.infrastructure.app.services
+{
+    using configuration;
+    using results;
+
+    /// <summary>
+    /// The automagic uninstaller service
+    /// </summary>
+    public interface IAutomaticUninstallerService
+    {
+        /// <summary>
+        /// Runs to remove a application from the registry.
+        /// </summary>
+        /// <param name="packageResult">The package result.</param>
+        /// <param name="config">The configuration.</param>
+        void run(PackageResult packageResult, ChocolateyConfiguration config);
+    }
+}
